requirements.txt
attrs==21.2.0
blis==0.7.8
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cymem==2.0.6
deprecation==2.1.0
dkpro-cassis==0.7.2
filelock==3.7.1
Flask==2.2.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
more-itertools==8.12.0
murmurhash==1.0.7
numpy==1.23.1
packaging==21.3
pathy==0.6.2
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.9
python-crfsuite==0.9.8
requests==2.28.1
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==5.2.1
sortedcontainers==2.4.0
tabulate==0.8.10
threadpoolctl==3.1.0
toposort==1.7
tqdm==4.64.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
wasabi==0.10.1
Werkzeug==2.2.1
zipp==3.8.1
